media: ov6211: Adds support for OV6211
OV6211 is a black and white 160 kpixel (400x400) camera sensor from OmniVision.
It supports 400x400, 200x200 and 100x100 image sizes. With careful register
programming it is possible to have non-square sizes as well. The output formats
are 8/10-bit RAW, this driver is only using the former. OV6211 supports a
single lane MIPI transmitter interface with up to 600 Mbps data rate.